requirements.txt
Django==2.1.9
django-crispy-forms==1.7.2
django-filter==2.1.0
mysqlclient==1.4.2.post1
pytz==2019.1
